+++ spring-cloud-starter-tencent-polaris-config/src/main/java/com/tencent/cloud/polaris/config/adapter/PolarisConfigFileLocator.java
@@ -149,10 +149,12 @@ public class PolarisConfigFileLocator implements PropertySourceLocator {
 			}
 			internalConfigFiles.add(new DefaultConfigFileMetadata(namespace, serviceName, "application-" + profile + ".properties"));
 			internalConfigFiles.add(new DefaultConfigFileMetadata(namespace, serviceName, "application-" + profile + ".yml"));
+			internalConfigFiles.add(new DefaultConfigFileMetadata(namespace, serviceName, "application-" + profile + ".yaml"));
 		}
 		// build default config properties files.
 		internalConfigFiles.add(new DefaultConfigFileMetadata(namespace, serviceName, "application.properties"));
 		internalConfigFiles.add(new DefaultConfigFileMetadata(namespace, serviceName, "application.yml"));
+		internalConfigFiles.add(new DefaultConfigFileMetadata(namespace, serviceName, "application.yaml"));
 	}
 
 	private void buildInternalBootstrapConfigFiles(
@@ -163,10 +165,12 @@ public class PolarisConfigFileLocator implements PropertySourceLocator {
 			}
 			internalConfigFiles.add(new DefaultConfigFileMetadata(namespace, serviceName, "bootstrap-" + profile + ".properties"));
 			internalConfigFiles.add(new DefaultConfigFileMetadata(namespace, serviceName, "bootstrap-" + profile + ".yml"));
+			internalConfigFiles.add(new DefaultConfigFileMetadata(namespace, serviceName, "bootstrap-" + profile + ".yaml"));
 		}
 		// build default config properties files.
 		internalConfigFiles.add(new DefaultConfigFileMetadata(namespace, serviceName, "bootstrap.properties"));
 		internalConfigFiles.add(new DefaultConfigFileMetadata(namespace, serviceName, "bootstrap.yml"));
+		internalConfigFiles.add(new DefaultConfigFileMetadata(namespace, serviceName, "bootstrap.yaml"));
 	}
 
 	private void initCustomPolarisConfigFiles(CompositePropertySource compositePropertySource,