+++ spring-cloud-starter-tencent-polaris-discovery/src/test/java/com/tencent/cloud/polaris/discovery/reactive/PolarisReactiveDiscoveryClientConfigurationTest.java
@@ -71,6 +71,11 @@ public class PolarisReactiveDiscoveryClientConfigurationTest {
 				.hasSingleBean(PolarisReactiveDiscoveryClient.class));
 	}
 
+	@Test
+	public void shouldWorkWithDefaults() {
+		contextRunner.run(context -> assertThat(context).hasBean("polarisReactiveDiscoveryClientHealthIndicator"));
+	}
+
 	@Configuration
 	@EnableAutoConfiguration
 	static class PolarisReactiveDiscoveryClientConfiguration {