+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # Change Log
 ---
 
+- feature: support reactive discovery client health indicator.
+- feature: Enhance default configuration to support `application*.yaml` and `bootstrap*.yaml`.